backend/app/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Ping Lin", email = "pingli88@gmail.com"},
]
dependencies = [
    "pyyaml==6.0",
]
requires-python = ">=3.10"
license = {text = "Apache-2.0"}

[project.urls]
Homepage = ""
[tool.pdm]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
